extract audio and storage helpers
